set service of og in using time instead of setting time
